requirements.txt
torch>=2.2.0
transformers>=4.41.0
datasets>=2.20.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
tqdm
pyyaml
